refactor(BE): 서버 전체 구조, 의존성 변경 (#58)
