+++ server/src/app.module.ts
@@ -5,14 +5,13 @@ import { AppService } from './app.service';
 import { ConfigModule, ConfigService } from '@nestjs/config';
 import { User } from './domain/user/user.entity';
 import { Post } from './domain/post/post.entity';
-import { AuthController } from './auth/auth.controller';
-import { AuthService } from './auth/auth.service';
-import { AuthModule } from './auth/auth.module';
-import { HttpModule } from '@nestjs/axios';
+import { AuthModule } from './domain/auth/auth.module';
 
 @Module({
   imports: [
-    ConfigModule.forRoot(),
+    ConfigModule.forRoot({
+      isGlobal: true,
+    }),
     TypeOrmModule.forRootAsync({
       imports: [ConfigModule],
       inject: [ConfigService],
@@ -28,9 +27,8 @@ import { HttpModule } from '@nestjs/axios';
       }),
     }),
     AuthModule,
-    HttpModule,
   ],
-  controllers: [AppController, AuthController],
-  providers: [AppService, AuthService],
+  controllers: [AppController],
+  providers: [AppService],
 })
 export class AppModule {}
