--- server/src/auth/auth.module.ts
@@ -1,20 +1,0 @@
-import { HttpModule } from '@nestjs/axios';
-import { Module } from '@nestjs/common';
-import { ConfigService } from '@nestjs/config';
-import { JwtModule } from '@nestjs/jwt';
-import { AccessTokenStrategy } from './access-token.strategy';
-import { AuthController } from './auth.controller';
-import { AuthService } from './auth.service';
-import { RefreshTokenStrategy } from './refresh-token.strategy';
-
-@Module({
-  imports: [HttpModule, JwtModule.register({})],
-  providers: [
-    ConfigService,
-    AuthService,
-    AccessTokenStrategy,
-    RefreshTokenStrategy,
-  ],
-  controllers: [AuthController],
-})
-export class AuthModule {}
