+++ server/src/domain/auth/access-token.strategy.ts
