+++ server/src/domain/auth/auth.module.ts
@@ -1,0 +1,20 @@
+import { Module } from '@nestjs/common';
+import { AuthController } from './auth.controller';
+import { AuthService } from './auth.service';
+import { JwtModule } from '@nestjs/jwt';
+import { HttpModule } from '@nestjs/axios';
+import { TypeOrmModule } from '@nestjs/typeorm';
+import { User } from '../user/user.entity';
+import { AccessTokenStrategy } from './access-token.strategy';
+import { RefreshTokenStrategy } from './refresh-token.strategy';
+
+@Module({
+  imports: [
+    TypeOrmModule.forFeature([User]),
+    HttpModule,
+    JwtModule.register({}),
+  ],
+  providers: [AuthService, AccessTokenStrategy, RefreshTokenStrategy],
+  controllers: [AuthController],
+})
+export class AuthModule {}
